Add brightness adjust
Added a simple brightness adjuster to boost the brightness of imported textures, but since it just naively works on the palette, and features no contrast adjustment, it does not produce good results. A better solution would be to adjust the brightness and contrast of each texture properly, instead of simply brightening pixels in the palette.